src/plugins/nonebot_plugin_datastore/db.py
from typing import AsyncGenerator

from nonebot import get_driver
from sqlalchemy.ext.asyncio import create_async_engine
from sqlmodel import SQLModel
from sqlmodel.ext.asyncio.session import AsyncSession

from .config import plugin_config

engine = create_async_engine(plugin_config.database_url, echo=True)


@get_driver().on_startup
async def init_db():
    async with engine.begin() as conn:
        await conn.run_sync(SQLModel.metadata.create_all)


async def get_session() -> AsyncGenerator[AsyncSession, None]:
    async with AsyncSession(engine) as session:
        yield session
