src/plugins/nonebot_plugin_datastore/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-datastore"
version = "0.0.1"
description = "适用于 Nonebot2 的数据存储插件"
authors = ["hemengyang <hmy0119@gmail.com>"]
license = "MIT"
packages = [
    { include = "nonebot_plugin_datastore/*.py", from = ".." }
]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-plugin-localstore = "^0.1.0"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
